requirements.txt
streamlit>=1.28.0
weaviate-client>=4.0.0
sentence-transformers>=2.2.0
httpx>=0.24.0
python-dotenv>=1.0.0
Pillow>=9.0.0
numpy>=1.21.0
